[REFACTOR] Changed module structure for external loaders
In the future, external_wrappers.py will be deprecated in favour of the higher
level wrappers module which dynamically loads available readers
